# Allow specialization of Data.Graph functions